Wizard: organize steps components into the "components" folders
